requirements.txt
secure-smtplib
bs4
python-dotenv